--- src/aics-image-viewer/shared/enums.ts
@@ -1,17 +1,0 @@
-export enum ViewMode {
-  threeD = "3D",
-  xy = "XY",
-  xz = "XZ",
-  yz = "YZ",
-}
-
-export enum RenderMode {
-  volumetric = "volumetric",
-  maxProject = "maxproject",
-  pathTrace = "pathtrace",
-}
-
-export enum ImageType {
-  segmentedCell = "cell",
-  fullField = "fov",
-}
