+++ src/aics-image-viewer/shared/utils/test/parseUrl.test.ts
@@ -2,9 +2,9 @@ import { describe, expect, it } from "@jest/globals";
 
 import { parseKeyValueList } from "../../../state/deserialize";
 import { DEFAULT_TEST_VIEWER_CHANNEL_SETTING } from "../../../state/test/test_data";
+import { ViewMode } from "../../../state/types";
 import type { ChannelState, ViewerChannelStateParams, ViewerState } from "../../../state/types";
 import { getDefaultChannelState, getDefaultViewerState } from "../../constants";
-import { ViewMode } from "../../enums";
 import { parseViewerUrlParams, serializeViewerUrlParams } from "../parseUrl";
 import type { ViewerChannelSetting } from "../viewerChannelSettings";
