+++ src/aics-image-viewer/state/reset.ts
@@ -6,11 +6,10 @@ import {
   getDefaultChannelState,
   getDefaultViewerState,
 } from "../shared/constants";
-import { ViewMode } from "../shared/enums";
 import type { ViewerChannelSettings } from "../shared/utils/viewerChannelSettings";
 import { getEnabledChannelIndices, initializeOneChannelSetting } from "../shared/utils/viewerState";
-import type { ChannelState, ViewerState } from "../state/types";
 import type { ViewerStore } from "./store";
+import { type ChannelState, type ViewerState, ViewMode } from "./types";
 import { validateState } from "./util";
 
 export type ResetStateActions = {
