+++ src/index.ts
@@ -8,9 +8,7 @@ export type {
   ViewerChannelGroup,
   ViewerChannelSetting,
 } from "./aics-image-viewer/shared/utils/viewerChannelSettings";
-export type { ViewerState } from "./aics-image-viewer/state/types";
-
-export { ViewMode, RenderMode, ImageType } from "./aics-image-viewer/shared/enums";
+export { type ViewerState, ViewMode, RenderMode, ImageType } from "./aics-image-viewer/state/types";
 
 export type { AppProps } from "./aics-image-viewer/components/App/types";
 export type { RawArrayData, RawArrayInfo } from "@aics/vole-core";
